+++ src/main/java/com/atwoz/member/domain/member/dto/PhysicalProfileInfo.java
@@ -1,6 +1,6 @@
 package com.atwoz.member.domain.member.dto;
 
-import com.atwoz.member.domain.member.profile.physical_profile.YearManager;
+import com.atwoz.member.domain.member.profile.physical.YearManager;
 
 public record PhysicalProfileInfo(
         Integer birthYear,
