+++ src/main/java/com/atwoz/member/domain/member/profile/physical/YearManager.java
@@ -1,0 +1,6 @@
+package com.atwoz.member.domain.member.profile.physical;
+
+public interface YearManager {
+
+    int getCurrentYear();
+}
