--- src/main/java/com/atwoz/member/domain/member/profile/physical_profile/YearManager.java
@@ -1,6 +1,0 @@
-package com.atwoz.member.domain.member.profile.physical_profile;
-
-public interface YearManager {
-
-    int getCurrentYear();
-}
