+++ src/main/java/com/atwoz/mission/application/membermission/event/MemberMissionClearedEvent.java
@@ -1,6 +1,7 @@
 package com.atwoz.mission.application.membermission.event;
 
-import com.atwoz.member.domain.info.profile.body.Gender;
+
+import com.atwoz.member.domain.member.profile.physical.Gender;
 
 public record MemberMissionClearedEvent(
         Long memberId,
