+++ src/test/java/com/atwoz/mission/application/membermission/MemberMissionsServiceTest.java
@@ -1,6 +1,6 @@
 package com.atwoz.mission.application.membermission;
 
-import com.atwoz.member.domain.info.profile.body.Gender;
+import com.atwoz.member.domain.member.profile.physical.Gender;
 import com.atwoz.mission.domain.membermission.MemberMission;
 import com.atwoz.mission.domain.membermission.MemberMissions;
 import com.atwoz.mission.domain.membermission.MemberMissionsRepository;
